Оптимизация производительности: NPC видение, IK, пространственные запросы
- Пакетная обработка лучей видения NPC: вместо каждого NPC по очереди, все NPC сразу отправляют лучи, а движок обрабатывает их одним пакетом — снижает нагрузку на CPU
- Параллельная трассировка лучей видения через TaskScheduler на несколько потоков
- Кеширование результатов запросов ближайших объектов (GetNearest) с настраиваемым TTL и LRU-вытеснением
- SSE-оптимизация пересечения AABB в пространственном дереве
- Сферические пространственные запросы вместо приближения через коробки — NPC точнее определяют объекты вокруг
- Пакетная обработка IK-лучей (ноги NPC) — все лучи одного NPC обрабатываются за один вызов вместо трёх отдельных
- Throttling сканирования врагов для далёких NPC: далекие stalkers сканируют реже, ближние — как раньше
- LOD-множитель для бюджета памяти NPC: далёкие NPC тратят меньше ресурсов на запоминание объектов
- Обновление видения NPC только когда игра не на паузе
- Настройка физического timestep (fixed_step 0.01 -> 0.02) для лучшего баланса точности и производительности
- Синхронизация ph_step_time с fixed_step вместо захардкоженного значения
- Новые консольные команды для тонкой настройки производительности: npc_perf_vision_*, npc_perf_ik_*, npc_perf_memory_*, obj_nearest_cache_*, ps_scheduler_*
- Защита от краша при уничтожении NPC во время обработки пакета видения
- Кеширование указателей на статическую геометрию перед параллельной трассировкой для защиты от dangling pointers
- Удаление мёртвого кода в IK-системе (chose_best_plane, неиспользуемые поля VisionBatchEntry)

+++ src/xrGame/ai/stalker/ai_stalker_cover.cpp
@@ -36,11 +36,11 @@ namespace
 struct cover_find_eval_budget_guard
 {
     u32 m_saved{};
-    explicit cover_find_eval_budget_guard(u32 per_find_total)
+    explicit cover_find_eval_budget_guard(int per_find_total)
     {
         m_saved = g_npc_perf_cover_find_eval_budget_remaining;
         if (per_find_total > 0)
-            g_npc_perf_cover_find_eval_budget_remaining = per_find_total;
+            g_npc_perf_cover_find_eval_budget_remaining = static_cast<u32>(per_find_total);
         else
             g_npc_perf_cover_find_eval_budget_remaining = u32(-1);
     }
